+++ History.md
@@ -17,6 +17,7 @@
   * Preserve `BUNDLE_GEMFILE` env var when using `prune_bundler` (#1893)
   * Send 408 request timeout even when queue requests is disabled (#2119)
   * Rescue IO::WaitReadable instead of EAGAIN for blocking read (#2121)
+  * Replace deprecated `Bundler.with_clean_env` with `with_unbundled_env` (#2120)
 
 * Refactor
   * Remove unused loader argument from Plugin initializer (#2095)